Refactore RecipeDetailedDTO
